Merge MPAS-Dev::develop into release-stable (#8)
Merge latest changes from upstream repository.  Works with mpas-jedi branch of the same name.  See that corresponding PR for more information on mpas-bundle behavior.

There are a lot of commits in this PR due to it having been a long time since such a merge was conducted.  The conflicts between the two branches are dealt with in commits 5beb733 and after.  Some small code changes from the early days of MPAS-JEDI development is removed, because it no longer serves a purpose.

After this PR is merged, `release-stable` will be nearly up to date with MPAS-Dev::develop with the following exceptions for code that is only in the JCSDA-internal fork:
- convective/cloud diagnostics for PANDA-C
- 2-stream I/O with a static file shared across all dates
- adjoint of some halo communications used for static B
- bug fixes in core_init_atmosphere for initializing `rho_zz` and `xice`
- optional cmake build process and dependence on the `jedi-cmake` repository

Closes #7